Update login & registration screen
