+++ .gradle/config.properties
@@ -1,2 +1,2 @@
-#Mon Mar 18 15:54:24 AEDT 2024
-java.home=C\:\\Program Files\\Android\\Android Studio\\jbr
+#Tue Mar 19 21:15:14 AEDT 2024
+java.home=/Applications/Android Studio.app/Contents/jbr/Contents/Home
