+++ .idea/deploymentTargetDropDown.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="deploymentTargetDropDown">
     <value>
-      <entry key="app">
+      <entry key="Unnamed">
         <State />
       </entry>
     </value>
