+++ app/.gitignore
@@ -1,1 +1,3 @@
 /build
+local.properties
+.idea/*
