+++ app/src/main/java/com/example/homescreen/MainActivity.kt
@@ -62,250 +62,7 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-//            LoginScreen({ _, _ ->})
-//            RegistrationScreen({ _, _, _, _, _, _, _ ->}, {->})
-//            HomeScreen()
-            NutritionTracker()
-        }
-    }
-}
-
-@Composable
-fun LoginScreen(onLoginClicked: (String, String) -> Unit) {
-    var email by rememberSaveable { mutableStateOf("") }
-    var password by rememberSaveable { mutableStateOf("") }
-    var isPasswordVisible by remember { mutableStateOf(false) }
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = "Login to Your Account",
-            style = MaterialTheme.typography.headlineMedium,
-            modifier = Modifier.padding(bottom = 32.dp) // Add some space below the title
-        )
-        TextField(
-            value = email,
-            onValueChange = { email = it },
-            label = { Text("Email") },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        TextField(
-            value = password,
-            onValueChange = { password = it },
-            label = { Text("Password") },
-            visualTransformation = if (isPasswordVisible) VisualTransformation.None else PasswordVisualTransformation(),
-            trailingIcon = {
-                val icon = if (isPasswordVisible) painterResource(id = R.drawable.eye) else painterResource(id = R.drawable.hidden)
-                IconButton(onClick = { isPasswordVisible = !isPasswordVisible }) {
-                    Icon(
-                        painter = icon,
-                        contentDescription = "Show or hide password",
-                        modifier = Modifier.height(22.dp)
-                    )
-                }
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(16.dp))
-        Button(
-            onClick = { onLoginClicked(email, password) },
-            modifier = Modifier
-                .height(46.dp)
-                .width(190.dp)
-        ) {
-            Text("Login")
-        }
-        Spacer(modifier = Modifier.height(6.dp))
-        TextButton(onClick = { /* Handle forgot password */ }) {
-            Text("Forgot Password?")
-        }
-        Row {
-            TextButton(onClick = { /* Navigate to registration screen */ },
-                ) {
-                Text("Don't have an account? Sign Up!")
-            }
-        }
-        // Google Sign-In Button
-        Button(
-            onClick = { /* Implement Google Sign-In logic here */ },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp),
-            colors = ButtonDefaults.buttonColors(Color.Transparent),
-            contentPadding = PaddingValues(0.dp)
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.android_light_rd),
-                contentDescription = "Sign in with Google",
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp)
-            )
-        }
-    }
-}
-
-@RequiresApi(0)
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun RegistrationScreen(
-    onRegisterClicked: (String, String, String, String, String, String, Date) -> Unit,
-    onNavigateToLogin: () -> Unit // Navigate back to login screen
-) {
-    var firstName by rememberSaveable { mutableStateOf("") }
-    var lastName by rememberSaveable { mutableStateOf("") }
-    var email by rememberSaveable { mutableStateOf("") }
-    var password by rememberSaveable { mutableStateOf("") }
-    var isPasswordVisible by rememberSaveable { mutableStateOf(false) }
-    val genders = listOf("Male", "Female")
-    var isExpanded by rememberSaveable { mutableStateOf(false) }
-    var selectedGender by rememberSaveable { mutableStateOf(genders[0]) }
-    var phone by rememberSaveable { mutableStateOf("") }
-    val calendar = Calendar.getInstance()
-    calendar.set(2024, 0, 1) // month (0) is January
-    val datePickerState = rememberDatePickerState(
-        initialSelectedDateMillis = Instant.now().toEpochMilli()
-    )
-    var showDatePicker by rememberSaveable { mutableStateOf(false) }
-    var selectedDate by rememberSaveable { mutableStateOf(calendar.timeInMillis) }
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = "Create Your Account",
-            style = MaterialTheme.typography.headlineMedium,
-            modifier = Modifier.padding(bottom = 24.dp)
-        )
-        TextField(
-            value = firstName,
-            onValueChange = { firstName = it },
-            label = { Text("First Name") },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        TextField(
-            value = lastName,
-            onValueChange = { lastName = it },
-            label = { Text("Last Name") },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        TextField(
-            value = email,
-            onValueChange = { email = it },
-            label = { Text("Email") },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        TextField(
-            value = password,
-            onValueChange = { password = it },
-            label = { Text("Password") },
-            visualTransformation = if (isPasswordVisible) VisualTransformation.None else PasswordVisualTransformation(),
-            trailingIcon = {
-                val icon = if (isPasswordVisible) painterResource(id = R.drawable.eye) else painterResource(id = R.drawable.hidden)
-                IconButton(onClick = { isPasswordVisible = !isPasswordVisible }) {
-                    Icon(
-                        painter = icon,
-                        contentDescription = "Show or hide password",
-                        modifier = Modifier.height(22.dp)
-                    )
-                }
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        ExposedDropdownMenuBox(expanded = isExpanded, onExpandedChange = { isExpanded = it }) {
-            TextField(
-                modifier = Modifier
-                    .menuAnchor()
-                    .fillMaxWidth()
-                    .focusProperties {
-                        canFocus = false
-                    }
-                    .padding(bottom = 8.dp),
-                readOnly = true,
-                value = selectedGender,
-                onValueChange = {},
-                label = { Text("Gender") },
-                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = isExpanded) }
-            )
-            ExposedDropdownMenu(
-                expanded = isExpanded,
-                onDismissRequest = { isExpanded = false })
-            {
-                genders.forEach { selectionOption ->
-                    DropdownMenuItem(
-                        text = { Text(selectionOption) },
-                        onClick = {
-                            selectedGender = selectionOption
-                            isExpanded = false
-                        },
-                        contentPadding = ExposedDropdownMenuDefaults.ItemContentPadding,
-                    )
-                }
-            }
-        }
-        TextField(
-            value = phone,
-            onValueChange = { phone = it },
-            label = { Text("Phone Number") },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            OutlinedButton(
-                onClick = { showDatePicker = true },
-                modifier = Modifier
-                    .weight(1f)
-                    .height(46.dp)
-            ) { Text(text = "Enter Date of Birth") }
-            Spacer(modifier = Modifier.width(12.dp))
-            val formatter = SimpleDateFormat("dd/MM/yyyy", Locale.ROOT)
-            Text(
-                text = "${formatter.format(Date(selectedDate))}",
-                modifier = Modifier.weight(1f)
-            )
-        }
-        if (showDatePicker) {
-            DatePickerDialog(
-                onDismissRequest = { showDatePicker = false },
-                confirmButton = {
-                    TextButton(onClick = {
-                        showDatePicker = false
-                        selectedDate = datePickerState.selectedDateMillis!!
-                    }) {Text(text = "OK") }
-                },
-                dismissButton = {
-                    TextButton(onClick = { showDatePicker = false }) { Text(text = "Cancel") }
-                }
-            ) { DatePicker(state = datePickerState) }
-        }
-        Spacer(modifier = Modifier.height(22.dp))
-        Button(
-            onClick = { onRegisterClicked(firstName, lastName, email, password, selectedGender, phone, Date(selectedDate)) },
-            modifier = Modifier
-                .height(46.dp)
-                .width(190.dp)
-        ) { Text("Register") }
-        Row(
-            modifier = Modifier.padding(top = 16.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text("Already have an account? ")
-            TextButton(onClick = onNavigateToLogin) {Text("Login here!") }
+            HomeScreen()
         }
     }
 }
@@ -361,8 +118,6 @@ fun TabContent(title: String) {
 @Composable
 fun HomeScreenPreview() {
     HomeScreenTheme {
-//        LoginScreen({ _, _ ->})
-//        RegistrationScreen({ _, _, _, _, _, _, _ ->}, {->})
         HomeScreen()
     }
 }
